feat(playground): move ddl to separate component
