+++ packages/playground/src/components/message/response-message-tool.tsx
@@ -1,30 +1,11 @@
-import {
-	CheckIcon,
-	ChevronsLeftRightIcon,
-	ChevronsRightLeftIcon,
-	CirclePause,
-	HammerIcon,
-	MoreHorizontalIcon,
-	PanelRightCloseIcon,
-	PanelRightOpenIcon,
-	TvMinimalIcon,
-	XCircleIcon,
-} from "lucide-react";
+import { CheckIcon, CirclePause, HammerIcon, XCircleIcon } from "lucide-react";
 import { observer } from "mobx-react-lite";
 import { useTranslation } from "@semoss/i18n";
-import {
-	Button,
-	cn,
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuSeparator,
-	DropdownMenuTrigger,
-	Spinner,
-} from "@semoss/ui/next";
+import { Button, cn, Spinner } from "@semoss/ui/next";
 import { useLoadingMessage } from "@/hooks";
 import type { ResponseMessageStore, ToolStore } from "@/stores";
 import { RoomInlineTool } from "../room";
+import { ResponseMessageToolMenu } from "./response-message-tool-menu";
 
 interface ResponseMessageToolProps {
 	/** Message to render */
@@ -272,104 +253,12 @@ export const ResponseMessageTool: React.FC<ResponseMessageToolProps> = observer(
 						</span>
 					)}
 					{toolState.actionType === "menu" && (
-						<DropdownMenu>
-							<DropdownMenuTrigger asChild>
-								{effectiveIsLarge ? (
-									<Button
-										type="button"
-										size="icon"
-										variant="ghost"
-										className="mr-2 shrink-0"
-										onClick={(e) => e.stopPropagation()}
-									>
-										<MoreHorizontalIcon className="size-4" />
-									</Button>
-								) : (
-									<button
-										type="button"
-										className="flex shrink-0 cursor-pointer items-center self-stretch rounded-r-lg px-4.5 hover:bg-accent"
-										onClick={(e) => e.stopPropagation()}
-									>
-										<MoreHorizontalIcon className="size-4 text-muted-foreground" />
-									</button>
-								)}
-							</DropdownMenuTrigger>
-							<DropdownMenuContent align="end">
-								<DropdownMenuItem
-									onClick={() => {
-										if (
-											tool.isOpen &&
-											tool.display === "inline"
-										) {
-											tool.closeTool();
-										} else {
-											tool.openTool("inline");
-										}
-									}}
-								>
-									{tool.isOpen &&
-									tool.display === "inline" ? (
-										<ChevronsRightLeftIcon />
-									) : (
-										<ChevronsLeftRightIcon />
-									)}
-									{tool.isOpen && tool.display === "inline"
-										? t("tool.collapse")
-										: t("tool.openInline")}
-								</DropdownMenuItem>
-								<DropdownMenuItem
-									onClick={() => {
-										tool.openTool("inline");
-										tool.setIsExpanded(true);
-									}}
-								>
-									<TvMinimalIcon />
-									{t("tool.expand")}
-								</DropdownMenuItem>
-								<DropdownMenuItem
-									onClick={() => {
-										if (
-											tool.isOpen &&
-											tool.display === "sidebar"
-										) {
-											tool.closeTool();
-										} else {
-											tool.openTool("sidebar");
-										}
-									}}
-								>
-									{tool.isOpen &&
-									tool.display === "sidebar" ? (
-										<PanelRightOpenIcon />
-									) : (
-										<PanelRightCloseIcon />
-									)}
-									{tool.isOpen && tool.display === "sidebar"
-										? t("tool.closeInSidebar")
-										: t("tool.openInSidebar")}
-								</DropdownMenuItem>
-								{toolState.showCancelInMenu && (
-									<>
-										<DropdownMenuSeparator />
-										<DropdownMenuItem
-											variant="destructive"
-											onClick={() => {
-												message.saveToolExecution(
-													tool,
-													"",
-													"cancelled",
-													{},
-												);
-												tool.closeTool();
-											}}
-										>
-											<XCircleIcon />
-											{t("tool.cancel")}
-										</DropdownMenuItem>
-									</>
-								)}
-							</DropdownMenuContent>
-						</DropdownMenu>
+						<ResponseMessageToolMenu
+							message={message}
+							tool={tool}
+							effectiveIsLarge={effectiveIsLarge}
+							showCancelInMenu={toolState.showCancelInMenu}
+						/>
 					)}
 				</div>
